+++ tests/backend/rethinkdb/test_schema.py
